requirements.txt
py-cord-2.6.1
pyyaml
os
re
python-dotenv